backend/requirements.txt
asgiref>=3.3.2,<4
pytz==2020.1
sqlparse==0.3.1
autopep8==1.5.3
Babel==2.8.0
certifi==2020.6.20
dj-database-url==0.5.0
Django>=3.2,<4.0
djangorestframework>=3.12,<4.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==3.2.1
django-phonenumber-field==4.0.0
django-elasticsearch-dsl==8.0.0
elasticsearch>=8.0.0,<9.0.0
elasticsearch-dsl==8.15.0
gunicorn==20.0.4
phonenumberslite==8.12.6
pycodestyle==2.6.0
PyJWT==1.7.1
python-dateutil==2.8.1
six==1.15.0
toml==0.10.1
urllib3>=1.26.2,<2
whitenoise==5.1.0
psycopg2-binary